Separate migration from registrar init in Filebeat
